pyproject.toml
[project]
name = "microscale"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.0",
    "pyexiv2>=2.15.5",
]

[project.scripts]
microscale = "microscale.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.10"


[tool.uv]
package = true

[project.optional-dependencies]
dev = [
  "pytest",
  "mypy",
]
